refactor(client-lib): move Forges to forge-url
For use in rest-api in the get_artifact_info handler.
