--- client-lib/src/download/gitlab.rs
@@ -1,12 +1,0 @@
-use super::ForgesPathMethods;
-
-pub(super) struct GitlabForge;
-
-impl ForgesPathMethods for GitlabForge {
-    fn translate_download_to_release_path(&self, path: &str) -> String {
-        // GitLab URLs already use the release path directly.
-        // Unlike GitHub (/releases/download/ -> /releases/tag/),
-        // GitLab stores at /-/releases/tag/ which matches the URL structure.
-        path.to_string()
-    }
-}
