+++ client-lib/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod artifact_info;
 mod backend;
 pub mod constants;
 mod download;
