+++ file-auth/forge-url/src/lib.rs
@@ -1,6 +1,7 @@
 mod error;
 pub mod fileserver;
 mod forge_info;
+pub mod forges;
 pub mod github;
 pub mod gitlab;
 mod origin;
@@ -8,5 +9,6 @@ mod traits;
 
 pub use error::ForgeUrlError;
 pub use forge_info::ForgeInfo;
+pub use forges::{Forges, get_forge};
 pub use origin::path_prefix_from_url;
-pub use traits::ForgeTrait;
+pub use traits::{ForgeTrait, ForgesPathMethods};
