+++ file-auth/forge-url/src/traits.rs
@@ -1,6 +1,9 @@
 use std::path::Path;
 
-use crate::ForgeUrlError;
+use crate::error::ForgeUrlError;
+use crate::path_prefix_from_url;
+use constants::INDEX_FILE;
+use url::Url;
 
 pub trait ForgeTrait
 where
@@ -17,4 +20,30 @@ where
     fn branch(&self) -> &str;
     fn file_path(&self) -> &Path;
     fn raw_url(&self) -> &url::Url;
+}
+
+/// Build repository paths from artifact *download* URLs.
+///
+/// Distinct from `ForgeTrait`, which parses signers/blob/raw URLs into repo
+/// info. Implementors only provide `translate_download_to_release_path`; the
+/// path builders are default methods.
+pub trait ForgesPathMethods {
+    fn construct_index_file_path(&self, file_url: &Url) -> Result<String, ForgeUrlError> {
+        self.construct_file_repo_path(file_url, INDEX_FILE)
+    }
+
+    fn construct_file_repo_path(
+        &self,
+        file_url: &Url,
+        filename: &str,
+    ) -> Result<String, ForgeUrlError> {
+        let prefix = path_prefix_from_url(file_url)?;
+        let path = file_url.path();
+        let path = path.strip_prefix('/').unwrap_or(path);
+        let dir_path = path.rsplit_once('/').map(|(dir, _)| dir).unwrap_or("");
+        let translated_path = self.translate_download_to_release_path(dir_path);
+        Ok(format!("{}/{}/{}", prefix, translated_path, filename))
+    }
+
+    fn translate_download_to_release_path(&self, path: &str) -> String;
 }
